+++ state/services.go
@@ -38,7 +38,7 @@ type BlockStore interface {
 // EvidencePool defines the EvidencePool interface used by the ConsensusState.
 // Get/Set/Commit
 type EvidencePool interface {
-	PendingEvidence(int64) []types.Evidence
+	PendingEvidence(uint32) []types.Evidence
 	AddEvidence(types.Evidence) error
 	Update(*types.Block, State)
 	// IsCommitted indicates if this evidence was already marked committed in another block.
@@ -48,7 +48,7 @@ type EvidencePool interface {
 // MockEvidencePool is an empty implementation of EvidencePool, useful for testing.
 type MockEvidencePool struct{}
 
-func (m MockEvidencePool) PendingEvidence(int64) []types.Evidence { return nil }
-func (m MockEvidencePool) AddEvidence(types.Evidence) error       { return nil }
-func (m MockEvidencePool) Update(*types.Block, State)             {}
-func (m MockEvidencePool) IsCommitted(types.Evidence) bool        { return false }
+func (m MockEvidencePool) PendingEvidence(uint32) []types.Evidence { return nil }
+func (m MockEvidencePool) AddEvidence(types.Evidence) error        { return nil }
+func (m MockEvidencePool) Update(*types.Block, State)              {}
+func (m MockEvidencePool) IsCommitted(types.Evidence) bool         { return false }
